Brand , Product, Category, User,

+++ src/main/java/com/ecommerce/WatchStore/Services/BrandService.java
@@ -2,7 +2,9 @@ package com.ecommerce.WatchStore.Services;
 
 import com.ecommerce.WatchStore.DTO.BrandDTO;
 import com.ecommerce.WatchStore.Entities.Brand;
+import com.ecommerce.WatchStore.Entities.Product;
 import com.ecommerce.WatchStore.Repositories.BrandRepository;
+import com.ecommerce.WatchStore.Repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,10 +15,12 @@ import java.util.Optional;
 public class BrandService {
         @Autowired
         private  BrandRepository brandRepository;
+        @Autowired
+        private ProductRepository productRepository;
         public List<Brand> getAllBrands(){
             return brandRepository.findAll();
         }
-        public Optional<Brand> getBrandById(Long id){
+        public Optional<Brand> getBrandById(int id){
             return brandRepository.findById(id);
         }
 
@@ -38,11 +42,13 @@ public class BrandService {
         return brandRepository.save(brand);
     }
 
-    public Brand updateBrand(Brand updatedBrand , Long id){
+    public Brand updateBrand(Brand updatedBrand , int id){
             Optional<Brand> existingBrandOptional = brandRepository.findById(updatedBrand.getIdBrand());
             if (existingBrandOptional.isPresent()){
                 Brand existingBrand = existingBrandOptional.get();
                 existingBrand.setName(updatedBrand.getName());
+                existingBrand.setUpdatedDt(updatedBrand.getUpdatedDt());
+                existingBrand.setUpdatedBy(updatedBrand.getUpdatedBy());
                 return brandRepository.save(existingBrand);
             }
            else
@@ -51,8 +57,15 @@ public class BrandService {
             }
             
         }
-        public void deleteBrandById(Long id){
-             brandRepository.deleteById( id);
+        public void deleteBrandById(int id){
+            List<Product> products = productRepository.findProductsByBrandId(id);
+            if(products.isEmpty())
+            {
+                brandRepository.deleteById( id);
+            }
+            else {
+                throw new RuntimeException("Không thể xóa thương hiệu khi còn sản phẩm thuộc thương hiệu này.");
+            }
         }
